### 2023-10-10 #5937 @editorialbot
[PRE REVIEW]: PyRolL - An Extensible OpenSource Framework for Rolling Simulation
**Submitting author:** @axtimhaus (<a href="http://orcid.org/0000-0002-8232-6877">Max Weiner</a>)
**Repository:** https://github.com/pyroll-project/pyroll-core
**Branch with paper.md** (empty if default branch): joss
**Version:** 2.1
