+++ paper.bib
@@ -128,7 +128,8 @@
   volume = {79},
   number = {4},
   pages = {287--296},
-  file = {/home/maxwe/Zotero/storage/4WBIDVSJ/Dixon2008.pdf}
+  file = {/home/maxwe/Zotero/storage/4WBIDVSJ/Dixon2008.pdf},
+  doi = {10.1002/srin.200806353},
 }
 
 @article{Dong2006,
@@ -445,7 +446,8 @@
   volume = {30},
   pages = {622--628},
   abstract = {Neuer Ansatz zur Abbildung der Geometrie von Kalibern 4 Parameter notwendig Algorithmus zum Einlesen der neuen Geometriebeschreibung auf herkömmlichen Wegen Beherrscht keine Einschnürung},
-  file = {/home/maxwe/Zotero/storage/NFSCFG6S/huang2014.pdf}
+  file = {/home/maxwe/Zotero/storage/NFSCFG6S/huang2014.pdf},
+  doi = {10.1016/j.rcim.2014.04.006},
 }
 
 @thesis{Jaeckel1991,
@@ -606,7 +608,8 @@
   volume = {35},
   number = {9},
   pages = {1100--1108},
-  file = {/home/maxwe/Zotero/storage/6KCDMP5T/Lehnert1995.pdf}
+  file = {/home/maxwe/Zotero/storage/6KCDMP5T/Lehnert1995.pdf},
+  doi = {10.2355/isijinternational.35.1100},
 }
 
 @article{Lehnert1995a,
@@ -754,7 +757,8 @@
   number = {7-8},
   pages = {912--928},
   abstract = {Reviewarticle zu Optimierungsstrategien und Ansätze beim Walzen allgemein},
-  file = {/home/maxwe/Zotero/storage/6J2REB3G/Oduguwa und Roy - 2006 - A review of rolling system design optimisation.pdf;/home/maxwe/Zotero/storage/ISAVWQ6B/Review_Rolling_System_Design_Optimisation_2006.pdf}
+  file = {/home/maxwe/Zotero/storage/6J2REB3G/Oduguwa und Roy - 2006 - A review of rolling system design optimisation.pdf;/home/maxwe/Zotero/storage/ISAVWQ6B/Review_Rolling_System_Design_Optimisation_2006.pdf},
+  doi = {10.1016/j.ijmachtools.2005.07.023}
 }
 
 @article{Overhagen2020,
@@ -767,7 +771,8 @@
   pages = {343--348},
   publisher = {{Walter de Gruyter GmbH}},
   doi = {10.1515/teme-2019-0130},
-  file = {/home/maxwe/Zotero/storage/2X3C34BQ/Overhagen2020a.pdf}
+  file = {/home/maxwe/Zotero/storage/2X3C34BQ/Overhagen2020a.pdf},
+  doi = {10.25518/esaform21.3987},
 }
 
 @article{Overhagen2020a,
@@ -824,7 +829,8 @@
   pages = {60--67},
   abstract = {Paper zur Dissertation von Klasen},
   langid = {ngerman},
-  file = {/home/maxwe/Zotero/storage/G6HKP7FB/steel_research_1988_PawelskiRaspKlasen_Breitung_beim_Warmwalzen_auf_geneigter_Bahn.PDF}
+  file = {/home/maxwe/Zotero/storage/G6HKP7FB/steel_research_1988_PawelskiRaspKlasen_Breitung_beim_Warmwalzen_auf_geneigter_Bahn.PDF},
+  doi = {10.1002/srin.198801607},
 }
 
 @book{Pawelski2000,
@@ -858,7 +864,8 @@
   volume = {36},
   number = {6},
   pages = {257--270},
[…]
