### 2023-10-11 #5937 @axtimhaus
I have added the missing DOIs as mentioned in the list.

Regarding reviewers: Our field is really small, especially those guys with experience in programming. Most that we know already gave us feedback on our work, or are cooperating with us in other unrelated or loosely related projects. Some others have definitely competing interests, since they are related to commercial simulation tools. We can, however, still provide the information for you to assess.
