### 2023-10-16 384b187
Update equivalent_ripped_groove.py according to suggestions by Max Weiner and Christoph Renzing
