### 2023-10-17 5bdc3e1
Update equivalent_ripped_groove.py according to suggestions Christoph Renzing
