### 2023-10-17 3fc34ba
Add equivalent ripped groove (#87)
* started equivalent_ripped_groove RP will continue

* finished equivalent_ripped_groove version_1

* finished equivalent_ripped_groove version_1

* Finalized groove and test file for the EquivalentRibbedGroove

* Update equivalent_ripped_groove.py according to suggestions by Max Weiner and Christoph Renzing

* Update equivalent_ripped_groove.py according to suggestions Christoph Renzing

---------

Co-authored-by: Christoph Renzing <christoph.renzing@imf.tu-freiberg.de>
Co-authored-by: Richard Pfeifer <richard.pfeifer@student.tu-freiberg.de>
