### 2023-10-18 56ead6b
added negative construction value check as well as tests
