+++ pyroll/core/grooves/generic_elongation.py
@@ -66,6 +66,11 @@ class GenericElongationGroove(GrooveBase, ReprMixin):
         :param classifiers: sequence of additional type classifiers
         """
 
+        mandatory_positive_or_zero = [r1, r2, r3, r4, alpha3, alpha4, indent, even_ground_width, flank_angle,
+                                      usable_width, ground_width, depth]
+        if any(value < 0 for value in mandatory_positive_or_zero):
+            raise ValueError("Groove arguments have to be positive.")
+
         try:
             if usable_width is None:
                 if np.isclose(depth, 0):
