### 2023-10-18 2291a0e
added CONTRIBUTING.rst (forked from pyiron project)
