+++ CONTRIBUTING.rst
@@ -1,0 +1,531 @@
+======================
+Contributing to PyRolL
+======================
+
+The following is a set of guidelines for contributing to PyRolL, which is
+hosted and maintained by the `PyRolL-Project Core Team`_
+on GitHub. These are mostly guidelines to facilitate an efficient
+development workflow, and not necessarily rules. Use your best judgment,
+and feel free to propose changes even to this document in a pull request.
+
+You can find all the PyRolL packages at our `github page`_ .
+To create pull requests, you will need to become part of the
+PyRolL organization. Please email us if you would like to join.
+
+Wait I don't want to read this; I just have a quick question/bugfix!
+====================================================================
+
+1. Check out our `FAQ page`_; your question might already be answered there.
+2. If your question relates to a bug in PyRolL, please briefly search the `issues page`_ and open a new labeled issue if you don't see anything related to your question there.
+3. You can also contact us on stackoverflow with the tag 'PyRolL'. There we will answer your questions. For bug fixes, please open an issue on github, as mentioned before.
+
+Table of Contents
+=================
+
+`License`_
+
+`What should I know before I get started?`_
+  * `PyRolL developer meetings`_
+
+..
+ * `The structure of PyRolL`_
+..
+ * `The principles of PyRolL`_
+
+
+`How can I contribute?`_
+  * `Reporting bugs`_
+  * `Suggesting enhancements`_
+  * `Your first code contribution`_
+  * `Pull requests`_
+
+`Styleguides`_
+  * `Git commit messages`_
+  * `Python styleguide`_
+  * `Documentation styleguide`_
+
+`Additional Notes`_
+  * `Issue and pull request labels`_
+  * `Build status`_
+  * `PyRolL release distribution`_
+  * `Building process for a release`_
+  * `GitHub Workflows`_
+  
+`Debugging`_
+  * `My job does not run on the queue`_
+
+License
+=======
+PyRolL is released as an open-source project under the BSD 3-Clause License.
+Code contributions should also be considered open-source.
+
+What should I know before I get started?
+========================================
+
+.. The structure of PyRolL
+.. -----------------------
+
+.. The principles of PyRolL
+.. ------------------------
+
+PyRolL developer meetings
+-------------------------
+If you are interested in discussing PyRolL's development, we encourage you to virtually
+participate in the PyRolL developer meeting Mondays at 15:00 german time (GMT+2).
+Contact us if you want to participate in the weekly virtual PyRolL developer meeting, become part of the PyRolL organization, and check the discussion page for details.
+
+How can I contribute?
+=====================
+
+Reporting bugs
+--------------
+
+    Note: If you find a closed issue that seems like it is the same
+    thing that you're experiencing, open a new issue and include a
+    link to the original issue in the body of your new one.
+
+**Before Submitting A Bug Report**
+
+Check if you can reproduce the problem in the latest version of PyRolL.
+Check the `FAQ page`_ for a list of common questions and problems.
+Briefly search the issues page for `bugs`_  to see if the problem has already
+been reported. If it has and the issue is still open, add a comment
+to the existing issue instead of opening a new one.
+
+**How Do I Submit A (Good) Bug Report?**
+
+Bugs are tracked as GitHub issues. You can create an issue on
+the PyRolL repository by including the following information:
+
+* Use a clear and descriptive title for the issue to identify the problem.
+* Describe the exact steps you took so we can reproduce the problem as closely as possible.
+* Provide sample code that causes the problem. Include code snippets as markdown code blocks.
+* Include information about the environment (OS, python version, how packages were installed) in which you were running PyRolL.
[…]
