### 2023-10-18 db6e977
added check for complex groove contour line.
