### 2023-10-19 fad55b3
Update pyroll/core/grooves/generic_elongation.py
Co-authored-by: Max Weiner <max.weiner@imf.tu-freiberg.de>
