### 2023-10-19 28ca47c
Added negative construction value check as well as tests (#101)
* added negative construction value check as well as tests

* Update pyroll/core/grooves/generic_elongation.py

Co-authored-by: Max Weiner <max.weiner@imf.tu-freiberg.de>

---------

Co-authored-by: Max Weiner <max.weiner@imf.tu-freiberg.de>
