### 2023-10-19 #5937 @Kevin-Mattheus-Moerman
@axtimhaus thanks for this submission. Could you clarify what aspects are under review here? The paper title says `PyRolL`, the repository listed above  here however points at `PyRolL-core`, but that repository shows there are 28 projects that depend on it, that all seem related: https://github.com/pyroll-project/pyroll-core/network/dependents. So, could you clarify what should be reviewed here please? Thanks
