### 2023-10-19 #5937 @axtimhaus
Please consider the project as a whole. PyRolL uses a highly modular approach to ensure extensibility and reuseability.

The core repository/package is the common basis for implementation of models and application logic. The other repositories contain plugins (modeling approaches) and extensions (application logic, UI). The user can install and import the different modules as needed to construct a simulation. See [an example in the docs](https://pyroll.readthedocs.io/en/latest/examples/step_by_step/index.html) on this.
