### 2023-10-19 #5937 @Kevin-Mattheus-Moerman
@axtimhaus thanks for the clarification.

As a recommendation (not a requirement just as one researcher to another) I would say that it would be great to extend your README with a bit of an explanation of what the software does. I would first briefly explain what rolling you are talking about, e.g. groove rolling. It would be great to have a graphic visualizing the rolling process. There are manual and automated groove rollers I suppose. Is your work relevant to both types? Next if you clarified what the typical inputs are, what the purpose of the simulation is, what its assumptions are (mention or refer to underlying theory), and what the typical outputs are, that would be great. Actually I would add as much visual information as possible. E.g. graphs or a table of typical outputs. I would also briefly mention what these outputs allow one to do? e.g. optimise/evaluate X for the process. At the moment as an outsider, it looks tricky to see what the software does.
If this was a FEM package the equivalent would be showing some meshes, the element types, the material model assumptions (e.g. linear elasticity/plasticity), some assumptions on boundary conditions, visualising some model predictions etc. Perhaps you can show some equivalent information? This doesn't have to be too technical/detailed. At the moment it just says "rolling simulation" and "groove rolling in elongation grooves" but it is unclear what this means.
I tried checking out the documentation, e.g. https://pyroll.readthedocs.io/en/latest/examples/basic.html but even that basic example doesn't say something like, "This is a example of the use of PyRolL for the evaluation of X and Y for this and this rolling process. Figure X shows the typical set-up for such a rolling process. This example helps to evaluate the rolling force, and the final shape of the.... " I feel that in many cases this type of detail is missing. The same for this step by step one: https://pyroll.readthedocs.io/en/latest/examples/step_by_step/index.html. There is no brief introduction to say what it is about. One has to already know all the physics involved and understand your code approach to dive in and see from the code what is going on. Hope I do not sound overly critical, this work looks interesting and well prepared. I just think more breadcrumbs and introduction texts and visuals of the processes would be great.

FYI your link [here](https://pyroll.readthedocs.io/en/latest/examples/basic.html) to the "Institute of Metal Forming" doesn't work.
