### 2023-10-26 pyroll-project/pyroll-core#110 @ChRen95
Clarify Intentions and Specifics of PyRolL in the Readme
Pay respect to suggestions of @Kevin-Mattheus-Moerman [here](https://github.com/openjournals/joss-reviews/issues/5937#issuecomment-1770799637).

 "I would say that it would be great to extend your README with a bit of an explanation of what the software does. I would first briefly explain what rolling you are talking about, e.g. groove rolling. It would be great to have a graphic visualizing the rolling process. There are manual and automated groove rollers I suppose. Is your work relevant to both types? Next if you clarified what the typical inputs are, what the purpose of the simulation is, what its assumptions are (mention or refer to underlying theory), and what the typical outputs are, that would be great. Actually I would add as much visual information as possible. E.g. graphs or a table of typical outputs. I would also briefly mention what these outputs allow one to do? e.g. optimise/evaluate X for the process. At the moment as an outsider, it looks tricky to see what the software does."
