### 2023-10-27 c17a9d0
add aim and scope
