### 2023-11-03 1c3015f
added more test and checks for orientation
