+++ pyroll/core/roll_pass/roll_pass.py
@@ -242,9 +242,16 @@ class RollPass(DiskElementUnit, DeformationUnit):
         ax.grid(lw=0.5)
 
         def oriented(geom):
-            if self.orientation != 0:
+            if isinstance(self.orientation, int) and self.orientation != 0:
                 return rotate(geom, self.orientation, (0, 0))
-            return geom
+            elif self.orientation == 0:
+                return geom
+            elif self.orientation.lower() in ["horizontal", "h", "antiy", "ay"]:
+                return geom
+            elif self.orientation.lower() in ["vertical", "v"]:
+                return rotate(geom, 90, (0, 0))
+            elif self.orientation.lower() in ["y"]:
+                return rotate(geom, 180, (0, 0))
 
         artists = []
 
