### 2023-11-03 bfde5c2
fixed bug in value check
