+++ pyroll/core/grooves/generic_elongation.py
@@ -68,7 +68,7 @@ class GenericElongationGroove(GrooveBase, ReprMixin):
 
         mandatory_positive_or_zero = [r1, r2, r3, r4, alpha3, alpha4, indent, even_ground_width, flank_angle,
                                       usable_width, ground_width, depth]
-        if any(value < 0 for value in mandatory_positive_or_zero):
+        if all(value is None or value > 0 for value in mandatory_positive_or_zero):
             raise ValueError("Groove arguments have to be non-negative.")
 
         try:
