### 2023-11-06 7d1722e
Update orientation Hook (#112)
* renamed thermal_capacity to specific_heat_capacity

* updated orientation hook

* added more test and checks for orientation
