### 2023-11-06 736da7b
updated if loop and updated test to new error message
