### 2023-11-06 8484479
Fix bug in mandatory positive or zero (#114)
* renamed thermal_capacity to specific_heat_capacity

* updated orientation hook

* fixed bug in value check

* updated if loop and updated test to new error message
