+++ pyroll/core/roll_pass/three_roll_pass.py
@@ -18,12 +18,12 @@ class ThreeRollPass(RollPass):
             return self._contour_lines
 
         shift = self.roll.groove.usable_width / 2 / np.sqrt(3) + self.gap / np.sqrt(3)
-        lower = rotate(translate(self.roll.contour_line, yoff=shift), angle=180, origin=(0, 0))
-        lower = LineString(lower.coords[::-1])  # get back coordinate order
-        right = rotate(lower, angle=120, origin=(0, 0))
-        left = rotate(lower, angle=-120, origin=(0, 0))
+        upper = translate(self.roll.contour_line, yoff=shift)
+        upper = LineString(upper.coords[::-1])  # get back coordinate order
+        right = rotate(upper, angle=-120, origin=(0, 0))
+        left = rotate(upper, angle=120, origin=(0, 0))
 
-        self._contour_lines = [left, lower, right]
+        self._contour_lines = [right, upper, left]
         return self._contour_lines
 
     @property
