### 2023-11-13 e4fddaf
Make Y Configuration the Default for 3RP (#116)
* change default to Y configuration

* change default in rp plot
