+++ pyroll/core/roll_pass/hookimpls/roll_pass.py
@@ -112,7 +112,7 @@ def height3(self):
         self.roll.groove.usable_width / 2,
         math.inf
     )
-    return -2 * usable_contour.bounds[1]
+    return 2 * usable_contour.bounds[3]
 
 
 @RollPass.volume
