### 2023-11-15 c61298c
Added new hooks to roll and roll_pass as well as corresponding hookimpls.
