### 2023-11-15 26066f3
updated implementation of surface_velocity
