+++ pyroll/core/roll_pass/hookimpls/roll_pass.py
@@ -140,10 +140,10 @@ def contact_area3(self: ThreeRollPass):
 @RollPass.velocity
 def velocity(self: RollPass):
     if self.roll.has_value("neutral_point"):
-        alpha = np.arcsin(-self.roll.neutral_point / self.roll.working_radius)
+        return self.roll.surface_velocity * np.cos(self.roll.neutral_angle)
     else:
-        alpha = 0
-    return self.roll.surface_velocity * np.cos(alpha)
+        return self.roll.surface_velocity * np.cos(0)
+
 
 
 @RollPass.duration
