### 2023-11-16 eec3ed8
updated surface velocity hook
