### 2023-11-16 f4367fa
updated velocity hook
