### 2023-11-21 #5937 @Kevin-Mattheus-Moerman
@ChRen95 @axtimhaus thanks for the initial stab at the README. I think there is still work to be done, also on the documentation, and adding visualisations would be great too.

I will proceed to look for reviewers to avoid further delay. If at this point you are able to help suggest reviewers, please mention their names or GitHub handles here. If you do, please leave out the `@` symbol so they are not tagged.
