### 2023-11-28 #5937 @philipcardiff
Hi @JorgeDeLosSantos, would you review this submission for JOSS? The submission under consideration is PyRolL - An Extensible OpenSource Framework for Rolling Simulation (https://github.com/pyroll-project/pyroll-core). Based on your GitHub experiences with the finite element method, I think you are a suitable reviewer.

If you are unfamiliar with JOSS, the review process is unique: it takes place in a GitHub issue, is open, and encourages author-reviewer-editor conversations. You can have a look at our [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) and [review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html) to get an idea of what is expected.

Philip
