### 2023-11-28 6d787f6
Update paper.md
Minor changes to the English and a comment on the availability of open-source metal forming codes.
