+++ paper.md
@@ -45,51 +45,51 @@ bibliography: paper.bib
 
 # Summary
 
-Groove rolling is one of the main process routes for production of metal long products, such as bars, beams, wire and rods.
-The industry is currently under heavy pressure to optimize their processes regarding energy consumption while maintaining or increasing product quality.
+Groove rolling is one of the main process routes for the production of metal long products, such as bars, beams, wire and rods.
+The industry is currently under heavy pressure to optimize its processes regarding energy consumption while maintaining or increasing product quality.
 The introduction of new materials and alloys challenges production and technology engineers.
-In the past decades, numerical simulation tools became an integral part of process development and maintenance.
+In the past decades, numerical simulation tools have become an integral part of process development and maintenance.
 
-PyRolL is an open-source, modular and extensible framework aiming at numerical simulation of groove rolling processes and accompanied processes.
-PyRolL Core serves as the basis for model and application development by defining needed data structures, solution algorithms and providing a versatile plugin system.
-Rolling processes are a quite complex issue, since a simulation has to regard mechanical behavior of workpiece and plant, as well as thermodynamic and chemical processes occurring within the workpiece.
-The plugin system enables modular simulation setup, were the user can choose from a growing library of state-of-the-art model approaches published in scientific literature.
+PyRolL is an open-source, modular and extensible framework aiming at the numerical simulation of groove rolling processes and accompanying processes.
+PyRolL Core serves as the basis for model and application development by defining needed data structures and solution algorithms and providing a versatile plugin system.
+Rolling processes are quite a complex issue since a simulation has to regard the mechanical behavior of the workpiece and plant, as well as thermodynamic and chemical processes occurring within the workpiece.
+The plugin system enables modular simulation setup, where the user can choose from a growing library of state-of-the-art model approaches published in scientific literature.
 Additional and new model approaches can be implemented as plugin packages and used just the same as officially provided ones.
-By this concept, the ecosystem can grow and thus avoid the need to implement the basic stuff every time anew, so the user or developer can concentrate on the actual focus of his work.
+By this concept, the ecosystem can grow and thus avoid the need to implement the basic stuff every time, so the user or developer can concentrate on the actual focus of his work.
 
 # Statement of need
 
 Established in the late 19th century, mathematical modelling of groove rolling and optimization of used grooves, known as groove or roll pass design, was investigated by a variety of authors.
-@Geuze1900, @Brovot1903 and @Mercader1924 gave a first set of rules for design of grooves as well as simple equations for calculation of material spread in groove rolling.
+@Geuze1900, @Brovot1903 and @Mercader1924 gave the first set of rules for the design of grooves as well as simple equations for the calculation of material spread in groove rolling.
 Further investigations regarding this topic were carried out by @Siebel1924, who introduced the equivalent rectangle approach to derive the material spread in groove rolling from an equivalent flat roll pass.
 This method has been constantly supplemented by various authors [@Lendl1948; @Lendl1948a; @Lendl1949; @Spittel1984; @Lee2000; @Dong2006; @Dong2008].
[…]
