### 2023-12-01 42bd065
Updated cast and reference functions
