### 2023-12-01 2101c29
updated neutral point calculation (#127)
* updated neutral point calculation

* fix sign

---------

Co-authored-by: Max Weiner <max.weiner@imf.tu-freiberg.de>
