### 2023-12-01 762f617
123 add waterbox unit (#125)
* fixed missing list in transport.py

* added cooling pipe as a new transport section

* Updated cast and reference functions
