### 2023-12-01 978da06
added cooling pipe to init
