+++ pyroll/core/transport/__init__.py
@@ -1,3 +1,4 @@
 from .transport import Transport
+from .cooling_pipe import CoolingPipe
 
 from . import hookimpls
