### 2023-12-04 bec8173
added equivalent radius for profile
