+++ pyroll/core/__init__.py
@@ -12,7 +12,7 @@ from .hooks import Hook, HookHost, HookFunction, root_hooks
 from .disk_elements import DiskElementUnit
 from .config import Config, config
 
-VERSION = "2.1.3"
+VERSION = "2.1.4"
 
 root_hooks.update(
     {
