### 2023-12-08 e443a94
included badge and removed deprecated notebook
