### 2023-12-14 e0c5e5a
added attributes for special values in the equivalent ribbed groove
