### 2023-12-14 565ceb8
removed negative implementation of entry angle
