+++ pyroll/core/roll_pass/hookimpls/roll_pass.py
@@ -225,4 +225,4 @@ def roll_power_3(self: ThreeRollPass):
 
 @RollPass.entry_angle
 def entry_angle(self: RollPass):
-    return -np.arcsin(self.roll.contact_length / self.roll.working_radius)
+    return np.arcsin(self.roll.contact_length / self.roll.working_radius)
