### 2023-12-14 3488541
updated thermal_capacity to specific_heat_capacity
