### 2023-12-15 fb64538
added constructor for hexagon profiles as well as tests and updated rotator
