### 2023-12-15 d93f497
Improve Readme (#111)
* add aim and scope

* move up docs
