### 2023-12-15 88da257
add from_dxf class method to spline groove and add ezdxf to dependencies
