+++ pyproject.toml
@@ -37,6 +37,7 @@ dependencies = [
     "numpy ~= 1.19",
     "scipy ~= 1.9",
     "shapely ~= 2.0",
+    "ezdxf"
 ]
 
 dynamic = ["version"]
