### 2023-12-18 3636a52
updated thermal_capacity to specific_heat_capacity (#145)
