### 2023-12-18 ebe4413
109 dxf input for grooves and rolls (#146)
* add from_dxf class method to spline groove and add ezdxf to dependencies

* added test
