### 2023-12-19 d3398e1
included entry point and angle
