### 2023-12-19 3ba33a4
removed negative implementation of entry angle (#140)
* removed negative implementation of entry angle

* included entry point and angle

* refactored

* fixed small error
