### 2023-12-20 29c2718
added forward slip and exit point as well as angle
