### 2023-12-20 cadb80d
added forward slip and exit point as well as angle
