+++ pyroll/core/__init__.py
@@ -19,16 +19,16 @@ root_hooks.extend(
         RollPass.roll_force,
         RollPass.Roll.roll_torque,
         RollPass.elongation_efficiency,
+        Unit.power,
+        Unit.OutProfile.cross_section,
+        Unit.OutProfile.classifiers,
+        Unit.OutProfile.strain,
+        Unit.OutProfile.length,
+        Unit.OutProfile.t,
         RollPass.OutProfile.filling_ratio,
         RollPass.OutProfile.cross_section_filling_ratio,
         RollPass.OutProfile.filling_error,
         RollPass.OutProfile.cross_section_error,
-        Unit.OutProfile.cross_section,
-        Unit.OutProfile.strain,
-        Unit.OutProfile.length,
-        Unit.OutProfile.classifiers,
-        Unit.OutProfile.t,
         PassSequence.log_elongation,
-        Unit.power,
     ]
 )
