### 2024-01-05 33523c4
change meaning of diagonal and height according to discussion
