### 2024-01-05 2f88a1a
remove hexagon rotator as equivalent with standard 3fold one
