### 2024-01-05 2ac5d6b
Added profile type hexagon (#143)
* added constructor for hexagon profiles as well as tests and updated rotator

* Improve Readme (#111)

* add aim and scope

* move up docs

* updated thermal_capacity to specific_heat_capacity (#145)

* 109 dxf input for grooves and rolls (#146)

* add from_dxf class method to spline groove and add ezdxf to dependencies

* added test

* included badge and removed deprecated notebook (#137)

* removed negative implementation of entry angle (#140)

* removed negative implementation of entry angle

* included entry point and angle

* refactored

* fixed small er
[…]
