### 2024-01-05 d9130e0
Root hooks ordered (#148)
* add root hook list

* upd test to current behavior of common spreading models

* reorder root hooks
