### 2024-01-08 6bd0a11
added tension elongation and updated calculation
