### 2024-01-10 #5937 @rboman
Hello @philipcardiff Happy new year too.
I already had a quick look to it before my vacations (installation from PyPI and running basic simulations). I have also read the documentation. I'll continue in a few days.
