### 2024-01-10 #6200 @editorialbot
[REVIEW]: PyRolL - An Extensible OpenSource Framework for Rolling Simulation
**Submitting author:** @axtimhaus (<a href="http://orcid.org/0000-0002-8232-6877">Max Weiner</a>)
**Repository:** https://github.com/pyroll-project/pyroll-core
**Branch with paper.md** (empty if default branch): joss
**Version:** v2.1.3
