### 2024-01-10 #5937 @philipcardiff
@rboman, @axtimhaus, we can continue the discussion at the review issue (https://github.com/openjournals/joss-reviews/issues/6200).
