+++ pyproject.toml
@@ -31,7 +31,7 @@ classifiers = [
     "Framework :: Sphinx",
     "Natural Language :: English",
 ]
-requires-python = ">=3.9"
+requires-python = ">=3.9, <4.0"
 
 dependencies = [
     "numpy ~= 1.19",
