### 2024-01-10 #6200 @philipcardiff
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.1098/rspa.1972.0025 is OK
> - 10.12688/f1000research.23224.2 is OK
> - 10.1016/j.commatsci.2005.01.009 is OK
> - 10.1002/srin.200806353 is OK
> - 10.2355/isijinternational.46.1458 is OK
> - 10.1016/S1005-8850(08)60065-1 is OK
> - 10.1179/030192304225011016 is OK
> - 10.1002/srin.200506068 is OK
> - 10.1002/srin.196705760 is OK
> - 10.1016/j.jmatprotec.2004.12.007 is OK
> - 10.1038/s41586-020-2649-2 is OK
> - 10.1016/j.rcim.2014.04.006 is OK
> - 10.1016/j.commatsci.2018.07.043 is OK
> - 10.1002/nme.1328 is OK
> - 10.1002/srin.200706280 is OK
> - 10.1007/s11665-008-9289-2 is OK
> - 10.1016/S0924-0136(00)00734-2 is OK
> - 10.1016/S0921-5093(01)01964-5 is OK
> - 10.1016/S0924-0136(02)00347-3 is OK
> - 10.2355/isijinternational.35.1100 is OK
> - 10.2355/isijinternational.35.1100 is OK
> - 10.2355/isijinternational.42.868 is OK
> - 10.1016/0924-0136(96)02307-2 is OK
> - 10.1016/j.ijmachtools.2005.07.023 is OK
> - 10.25518/esaform21.3987 is OK
> - 10.5281/ZENODO.6053272 is OK
> - 10.1002/srin.198801607 is OK
> - 10.1051/metal/193936060257 is OK
> - 10.1016/0890-6955(92)90022-9 is OK
> - 10.1016/j.proeng.2015.12.038 is OK
> - 10.1243/PIME_PROC_1961_175_043_02 is OK
> - 10.1016/j.proeng.2014.09.149 is OK
> - 10.1016/j.jmatprotec.2007.04.124 is OK
> - 10.1007/978-3-322-88346-9 is OK
> - 10.2355/isijinternational.40.65 is OK
>
> MISSING DOIs
>
> - 10.25518/esaform21.3987 may be a valid DOI for title: A Computational Method for Pass Design of the Four-Roll Rolling Process Forsizing of Round Sections
>
> INVALID DOIs
>
> - None

Hi @axtimhaus, can you check this possible missing DOI? Thanks.
