+++ paper.bib
@@ -793,7 +793,8 @@
   date = {2021},
   url = {https://popups.uliege.be/esaform21/index.php?id=3987},
   eventtitle = {{{ESAFORM}} 21},
-  file = {/home/maxwe/Zotero/storage/W3ATAKG5/Overhagen2021.pdf}
+  file = {/home/maxwe/Zotero/storage/W3ATAKG5/Overhagen2021.pdf},
+  doi = {10.25518/esaform21.3987}
 }
 
 @software{pandas,
