### 2024-01-11 aac398d
add ipython completion support
