### 2024-01-11 80fc9ea
add inscribed circle diameter hook and impls
