### 2024-01-12 6fa1a4d
Update three_roll_pass.py
Fixed typo
