+++ pyroll/core/roll_pass/three_roll_pass.py
@@ -13,7 +13,7 @@ class ThreeRollPass(RollPass):
     """Represents a roll pass with three working rolls and 3-fold symmetry."""
 
     inscribed_circle_diameter = Hook[float]()
-    """Diameter of incribed circle between roll barrels as alternative to roll gap definition."""
+    """Diameter of inscribed circle between roll barrels as alternative to roll gap definition."""
 
     @property
     def contour_lines(self) -> List[LineString]:
