### 2024-01-12 ddff7b9
add inscribed circle diameter hook and impls (#155)
* add inscribed circle diameter hook and impls

* Update three_roll_pass.py

Fixed typo

---------

Co-authored-by: Christoph Renzing <64863686+ChRen95@users.noreply.github.com>
