### 2024-01-12 bf911dc
added forward slip and exit point as well as angle (#156)
Co-authored-by: Christoph Renzing <christoph.renzing@imf.tu-freiberg.de>
