### 2024-01-12 1384c53
add label based indexing to pass sequence (#154)
* add label based indexing to pass sequence

* add ipython completion support
