### 2024-01-12 #6200 @rboman
Hi, I have posted my review as a MR (pyroll-docs) and an issue (pyroll-core).

Globally, I think that this piece of software is very interesting and deserves a paper in JOSS.
As the authors say in their document, this kind of numerical model is usually kept private in companies. Consequently making it open-source is a real gift for researchers (and companies). It is also very modular and built in a modern way. The code is clean. I also like the plugin system with hooks. I haven't had enough time to understand how it internally works but I am sure I'll have a closer look in the future for my own projects.

I have been very happy with the documentation although I saw that there is an issue which says that it should be completely rewritten. PyRolL is very easy to install with PyPI and it is also straightforward to make a quick simulation (I like the "create-input-py" command of the pyroll executable which gives a first basic test for the new user). The input and outputs are easily understandable and the generated HTML report is very convenient to analyse the results. I have also played with the example Jupyter notebooks (pyroll-examples). Everything works as expected.

Therefore, I have no "major changes" to request. I will be happy to accept the paper as soon as my comments in the issue have been resolved.
