### 2024-01-12 pyroll-project/pyroll-core#158 @axtimhaus
Thanks for your review. I understand the tenor of your suggestions and I basically agree with them. We will look to merge them as soon as possible!

They accompany with those of @philipcardiff in pyroll-project/pyroll-docs#8 for the docs.
