### 2024-01-12 #6200 @philipcardiff
For future reference: https://github.com/pyroll-project/pyroll-core/issues/158#issuecomment-1889612249
