### 2024-01-15 pyroll-project/pyroll-core#158 @axtimhaus
I have regarded your statements in 027b28b5.

Especially for the FEM part:
Yes, we generally target on fast models and do not try to compete with the depth and accuracy of FEM, but they are currently widely used and something like the top dog. We are always request to compare with them.
I shortened the paragraph and focused more on the difference to us to clarify that. Also I stated, that including a FE kernel is generally possible and planned, but currently low priority.
