### 2024-01-16 40cde27
add contact stresses to deformation unit
