+++ pyroll/core/roll_pass/deformation_unit.py
@@ -54,6 +54,12 @@ class DeformationUnit(Unit):
     deformation_resistance = Hook[float]()
     """Equivalent deformation resistance (mean flow stress increased by deformation efficiency)."""
 
+    contact_pressure = Hook[float]()
+    """Pressure acting on the contact area."""
+
+    contact_friction = Hook[float]()
+    """Friction stress acting on the contact area."""
+
     class Profile(Unit.Profile):
         """Represents a profile in context of a deformation unit."""
 
