### 2024-01-16 fdce11f
add contact geometry to deformation profiles
