### 2024-01-17 #6200 @axtimhaus
The latest release is 2.1.3 archived at https://doi.org/10.5281/zenodo.10495163

Thank you both for the review. It was pleasant and productive, wish it could be always like that.
