### 2024-01-18 #6200 @ChRen95
Just noticed, that we forgot to mention one valuable contributor which we want to thank. I just updated the paper.md.
