### 2024-01-19 b5ab3f2
Update JOSS paper
