+++ paper.md
@@ -45,7 +45,7 @@ bibliography: paper.bib
 
 # Summary
 
-Groove rolling is one of the main process routes for the production of metal long products, such as bars, beams, wires and rods.
+Groove rolling is one of the main process routes for the production of long metal products, such as bars, beams, wires, and rods.
 In this process, a workpiece is deformed by two or more rotating rolls to reduce the cross-section achieve a desired cross-section shape.
 The shape is determined by the contour of the rolls' surfaces, called the groove or calibre.
 The process is usually accomplished in multiple steps, each called a pass.
@@ -55,17 +55,17 @@ In the past decades, numerical simulation tools have become an integral part of 
 There are used to estimate forces and engine torques, as well as geometrical and microstructural evolution of the workpiece for a given pass sequence and input workpiece geometry and material.
 This knowledge is important to design or optimize a process, identify sources of errors, increase product quality or extend the product portfolio.
 
-PyRolL is an open-source, modular and extensible framework aiming at the numerical simulation of groove rolling processes and accompanying processes, such as heating, cooling and transportation.
+PyRolL is an open-source, modular, and extensible framework aiming at the numerical simulation of groove rolling processes and accompanying processes, such as heating, cooling, and transportation.
 PyRolL Core serves as the basis for model and application development by defining needed data structures and solution algorithms and providing a versatile plugin system.
 Rolling processes are quite a complex issue since a simulation has to regard the mechanical behavior of the workpiece and plant, as well as thermodynamic and chemical processes occurring within the workpiece.
 The plugin system enables modular simulation setup, where the user can choose from a growing library of state-of-the-art model approaches published in scientific literature that target specific aspects of the process such as material flow, stress state, thermal evolution, microstructure evolution or elastic tool response.
 Additional and new model approaches can be implemented as plugin packages and used just the same as officially provided ones.
-By this concept, the ecosystem can grow and thus avoid the need to implement the basic stuff every time, so the user or developer can concentrate on the actual focus of his work.
+By this concept, the ecosystem can grow and thus avoid the need to implement the basics every time, so the user or developer can concentrate on the actual focus of their work.
 
 # Statement of need
 
 Established in the late 19th century, mathematical modelling of groove rolling and optimization of used grooves, known as groove or roll pass design, was investigated by a variety of authors.
-@Geuze1900, @Brovot1903 and @Mercader1924 gave the first set of rules for the design of grooves as well as simple equations for the calculation of material spread in groove rolling.
+@Geuze1900, @Brovot1903, and @Mercader1924 gave the first set of rules for the design of grooves as well as simple equations for the calculation of material spread in groove rolling.
 Further investigations regarding this topic were carried out by @Siebel1924, who introduced the equivalent rectangle approach to derive the material spread in groove rolling from an equivalent flat roll pass.
 This method has been constantly supplemented by various authors [@Lendl1948; @Lendl1948a; @Lendl1949; @Spittel1984; @Lee2000; @Dong2006; @Dong2008].
 Furthermore empirical and analytical models have been developed, focusing on different aspects of groove rolling as spreading [@Tafel1925; @Siebel1924; @Ekelund1927; @Roux1939; @Marini1941; @Sparling1961; @Geleji1967; @Wusatowski1969; @Vater1972; @El-Nikhaily1979; @Klasen1984; @Angott1986; @Pawelski1988; @Dixon1996], longitudinal tension [@Treis1968; @Nikkila1977; @Voigtlaender1984; @Schulze1986; @Jaeckel1991; @Lommatzsch1991; @Shokhin2015], strain rate [@Hensel1985], contact area between work roll and rolled stock [@Zouhar1960; @Zouhar1966] and power and work needed [@Zouhar1960; @Zouhar1966; @Hensel1977; @Goldhahn1981; @Hensel1985].
[…]
