### 2024-01-19 #6200 @kyleniemeyer
Hi @axtimhaus, I noticed that the author lists differ between the paper and Zenodo archive. We typically ask that those match, unless for a particular reason. Can you fix that, or let us know a reason for the difference (if intentional)?

Also, I made some minor edits to the paper in https://github.com/pyroll-project/pyroll-core/pull/164. Can you review and merge?
