### 2024-01-22 9e41bb6
Update JOSS paper (#164)
* Update JOSS paper

* Update paper.md

changed to metal long products

* Update paper.md

---------

Co-authored-by: Christoph Renzing <64863686+ChRen95@users.noreply.github.com>
