+++ paper.md
@@ -46,19 +46,19 @@ bibliography: paper.bib
 # Summary
 
 Groove rolling is one of the main process routes for the production of metal long products, such as bars, beams, wires, and rods.
-In this process, a workpiece is deformed by two or more rotating rolls to reduce the cross-section achieve a desired cross-section shape.
+In this process, a workpiece is deformed by two or more rotating rolls to reduce the cross-section to achieve a desired cross-section shape.
 The shape is determined by the contour of the rolls' surfaces, called the groove or calibre.
 The process is usually accomplished in multiple steps, each called a pass.
 The industry is currently under heavy pressure to optimize its processes regarding energy consumption while maintaining or increasing product quality.
 The introduction of new materials and alloys challenges production and technology engineers.
 In the past decades, numerical simulation tools have become an integral part of process development and maintenance.
-There are used to estimate forces and engine torques, as well as geometrical and microstructural evolution of the workpiece for a given pass sequence and input workpiece geometry and material.
+They are used to estimate forces and engine torques, as well as geometrical and microstructural evolution of the workpiece for a given pass sequence and input workpiece geometry and material.
 This knowledge is important to design or optimize a process, identify sources of errors, increase product quality or extend the product portfolio.
 
 PyRolL is an open-source, modular, and extensible framework aiming at the numerical simulation of groove rolling processes and accompanying processes, such as heating, cooling, and transportation.
 PyRolL Core serves as the basis for model and application development by defining needed data structures and solution algorithms and providing a versatile plugin system.
 Rolling processes are quite a complex issue since a simulation has to regard the mechanical behavior of the workpiece and plant, as well as thermodynamic and chemical processes occurring within the workpiece.
-The plugin system enables modular simulation setup, where the user can choose from a growing library of state-of-the-art model approaches published in scientific literature that target specific aspects of the process such as material flow, stress state, thermal evolution, microstructure evolution or elastic tool response.
+The plugin system enables a modular simulation setup, where the user can choose from a growing library of state-of-the-art model approaches published in scientific literature that target specific aspects of the process such as material flow, stress state, thermal evolution, microstructure evolution or elastic tool response.
 Additional and new model approaches can be implemented as plugin packages and used just the same as officially provided ones.
 By this concept, the ecosystem can grow and thus avoid the need to implement the basics every time, so the user or developer can concentrate on the actual focus of their work.
 
@@ -80,7 +80,7 @@ The finite element theory is actively developed since the 1980s for use in groov
 It provides a general approach for complex problems in two or three dimensions and offers simulation results of high accuracy and depth.
 The main disadvantage is the high computational effort in solving the equation systems, increasing rapidly with the non-linearity of the model equations and the resolution of the solution space, typically in the range from hours to days.
 Analytical and empirical models, however, do not provide that high accuracy and depth, but are usually computable within fractions of seconds.
-So they are suitable for interactive design with quick feedback loops and numerical applications, which require large count of evaluations, such as optimization routines.
+Therefore they are suitable for interactive design with quick feedback loops and numerical applications, which require large count of evaluations, such as optimization routines.
 
 According to the authors' experience, it is not common in the field of rolling simulation to provide source code and input files alongside journal publications, so the work cannot be reproduced directly and easily.
 Noteable exceptions from this are @Alexander1972 and a textbook by @Pawelski2000.
@@ -89,7 +89,7 @@ In effect, every research project has to start from scratch to build up a simula
[…]
