### 2024-01-23 #6200 @axtimhaus
I have updated the Zenodo record. These were automatically filled from the Python metadata.
